Move version.py into package
