+++ MANIFEST.in
@@ -7,7 +7,6 @@ include LICENSE
 include MANIFEST.in
 include README.rst
 include setup.py
-include version.py
 recursive-include ci *.py
 recursive-include doc *
 recursive-include src *
