+++ src/hdf5plugin/_version.py
